major changes

+++ maps.R
@@ -8,9 +8,11 @@ library(moinput)
 
 # download file if necessary
 
-download.file("https://www.naturalearthdata.com/http//www.naturalearthdata.com/download/10m/cultural/ne_10m_admin_0_countries.zip", "ne_10m_admin_0_countries.zip")
-unzip("ne_10m_admin_0_countries.zip", exdir = "mapData")
-file.remove("ne_10m_admin_0_countries.zip")
+if(!file.exists("mapData")){
+  download.file("https://www.naturalearthdata.com/http//www.naturalearthdata.com/download/10m/cultural/ne_10m_admin_0_countries.zip", "ne_10m_admin_0_countries.zip")
+  unzip("ne_10m_admin_0_countries.zip", exdir = "mapData")
+  file.remove("ne_10m_admin_0_countries.zip")
+}
 
 ### simple chart ###
